some changes.
Intelisense error:
The type 'EventToCommand' from assembly 'GalaSoft.MvvmLight.Platform' is
built with an older version of the Blend SDK, and is not supported in a
Windows Presentation Framework 4 project.
Cannot add instance of type 'EventToCommand' to a collection of type
'TriggerActionCollection'. Only items of type 'T' are allowed.

http://stackoverflow.com/questions/32477728/mvvm-light-and-eventtocommand-gives-invalid-markup-in-vs2015
1. Uninstall nuget packages MvvmLightLibs and CommonServiceLocator.
2. Reinstall nuget packages MvvmLightLibs and CommonServiceLocator.
3. Goto $Solution Folder$\packages\MvvmLightLibs.5.3.0.0\lib\net45
4. Run gacutil -i System.Windows.Interactivity.dll

For more WPF examples:
https://github.com/Microsoft/WPF-Samples/